+++ .gitignore
@@ -22,6 +22,7 @@ classes/
 .vscode/
 bin/
 .classpath
+.factorypath
 .project
 
 # macos
