+++ src/main/resources/arcane_repository.mixins.json
@@ -9,7 +9,6 @@
     "ItemEntityMixin",
     "ItemMixin",
     "LecternBlockMixin",
-    "PlayerInventoryMixin",
     "accessor.ItemEntityAccessor",
     "accessor.ItemUsageContextAccessor",
     "accessor.LootTablesAccessor",
